components/redteam-runner/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "fabric-redteam-runner"
description = "SingleAxis Fabric — scheduled adversarial testing (Layer 1)."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.12"
authors = [{ name = "AI5Labs Research OPC Private Limited", email = "fabric@singleaxis.ai" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
  "httpx>=0.27,<1.0",
  "opentelemetry-api>=1.28,<2.0",
  "opentelemetry-sdk>=1.28,<2.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28,<2.0",
  "pydantic>=2.8,<3.0",
  "pyyaml>=6.0,<7.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
# Garak and PyRIT are heavyweight, still churning on minor versions,
# and not needed by unit tests — the runner falls back to stub drivers
# when they aren't installed. Include them via the ``suites`` extra on
# the runtime image.
suites = [
  "garak>=0.9.0.15; python_version>='3.12'",
  "pyrit>=0.5.0; python_version>='3.12'",
]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "respx>=0.21",
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "types-PyYAML>=6.0",
]

[project.scripts]
fabric-redteam-runner = "fabric_redteam_runner.__main__:main"

[tool.hatch.version]
path = "src/fabric_redteam_runner/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/fabric_redteam_runner"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
  "-ra",
  "--strict-config",
  "--strict-markers",
  "--cov=fabric_redteam_runner",
  "--cov-report=term-missing",
  "--cov-fail-under=80",
]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["fabric_redteam_runner"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "S", "C4", "PIE", "RET", "SIM", "PL", "RUF"]
ignore = ["PLR0913"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "PLR2004"]
# Typer requires `typer.Option(...)` in argument defaults; that's the
# whole ergonomic. S105 flags the "PASS" literal on the Verdict enum
# as a hardcoded password — false positive.
"src/fabric_redteam_runner/__main__.py" = ["B008"]
"src/fabric_redteam_runner/results.py" = ["S105"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

# Garak and PyRIT are optional extras; the runner falls back to stub
# drivers when they aren't installed. Silence "cannot find stub"
# errors so the default dev install (without the `suites` extra)
# still passes strict mypy.
[[tool.mypy.overrides]]
module = ["pyrit.*", "garak.*"]
ignore_missing_imports = true
